requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.28
joblib>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Optional dependencies
mlflow>=2.8.0
dvc>=3.0.0
optuna>=3.4.0
xgboost>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
